design corrections
